ai_proxy/requirements.txt
flask>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0